+++ MoveableImpl/src/com/myownb3/piranha/statemachine/impl/EvasionStateMachineConfigImpl.java
@@ -14,7 +14,7 @@ public class EvasionStateMachineConfigImpl implements EvasionStateMachineConfig 
     // Attributes of the different EvasionStateHandlerse
     private int postEvasionAngleAdjustStepWidth;
     private double returningMinDistance;
-    private int returningAngleIncMultiplier;
+    private double returningAngleIncMultiplier;
     private int passingDistance;
 
     // Attributes for Detector
@@ -24,13 +24,13 @@ public class EvasionStateMachineConfigImpl implements EvasionStateMachineConfig 
     private double evasionAngleInc;
     private double returningAngleMargin;
 
-    public EvasionStateMachineConfigImpl(int angleIncMultiplier, double minDistance, double angleMargin,
+    public EvasionStateMachineConfigImpl(double angleIncMultiplier, double minDistance, double angleMargin,
 	    int detectorReach, int detectorAngle, int evasionAngle, double evasionAngleInc) {
 	this(angleIncMultiplier, minDistance, angleMargin, detectorReach, 2 * detectorReach / 3, detectorAngle,
 		evasionAngle, evasionAngleInc);
     }
 
-    private EvasionStateMachineConfigImpl(int angleIncMultiplier, double minDistance, double angleMargin,
+    private EvasionStateMachineConfigImpl(double angleIncMultiplier, double minDistance, double angleMargin,
 	    int detectorReach, int passingDistance, int detectorAngle, int evasionAngle, double evasionAngleInc) {
 	this.postEvasionAngleAdjustStepWidth = 10; // Like this the movements are smoother
 	this.returningAngleIncMultiplier = angleIncMultiplier;
@@ -60,7 +60,7 @@ public class EvasionStateMachineConfigImpl implements EvasionStateMachineConfig 
     }
 
     @Override
-    public final int getReturningAngleIncMultiplier() {
+    public final double getReturningAngleIncMultiplier() {
 	return this.returningAngleIncMultiplier;
     }
